progress: Binary Serialization of AssetPack, Scene, and UIPage. Bytecode serialization of ScriptModule

+++ VolcaniCore/src/VolcaniCore/Core/Entrypoint.cpp
@@ -8,7 +8,7 @@ extern Application* CreateApplication(const CommandLineArgs& args);
 int main(int argc, char** argv) {
 	Application::SetCurrentDir();
 	Application::Init();
-	Application* app = CreateApplication(CommandLineArgs((uint32_t)argc, argv));
+	Application* app = CreateApplication(CommandLineArgs(argc, argv));
 	Application::Run();
 	Application::Close();
 }
